Time: 32 ms (71.77%), Space: 10.6 MB (62.58%) - LeetHub

+++ 923-3sum-with-multiplicity/923-3sum-with-multiplicity.cpp
@@ -1,6 +1,6 @@
 class Solution {
 private:
-    long mod = (int)1e9+7;
+    int mod = (int)1e9+7;
 public:
     int threeSumMulti(vector<int>& arr, int target) {
